[redirector] Separate app setup from routing specific functions from main.go
